+++ web/src/screens/CourseDetail.page.tsx
